requirements.txt
python-dotenv
requests
Pillow
browserforge[all]
duckduckgo_search
openai
genanki
markdown